app.py
import os
import logging
from flask import Flask
from flask_sqlalchemy import SQLAlchemy
from flask_socketio import SocketIO
from sqlalchemy.orm import DeclarativeBase
from werkzeug.middleware.proxy_fix import ProxyFix
from dotenv import load_dotenv
load_dotenv()

import os
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")


# Configure logging
logging.basicConfig(level=logging.DEBUG)

class Base(DeclarativeBase):
    pass

db = SQLAlchemy(model_class=Base)
socketio = SocketIO()

# create the app
app = Flask(__name__, static_folder='static', template_folder='templates')
socketio = SocketIO(app)
app.secret_key = os.environ.get("SESSION_SECRET", "dev-secret-key-change-in-production")
app.wsgi_app = ProxyFix(app.wsgi_app, x_proto=1, x_host=1)

# configure the database
app.config["SQLALCHEMY_DATABASE_URI"] = os.environ.get("DATABASE_URL", "sqlite:///legal_chatbot.db")
app.config["SQLALCHEMY_ENGINE_OPTIONS"] = {
    "pool_recycle": 300,
    "pool_pre_ping": True,
}

# initialize extensions
db.init_app(app)
socketio.init_app(app, cors_allowed_origins="*", async_mode='threading')

# Initialize the database and import models
with app.app_context():
    import models  # noqa: F401
    db.create_all()
    
    # Initialize sample data
    try:
        from initialize_data import initialize_all_sample_data
        initialize_all_sample_data()
    except Exception as e:
        logging.error(f"Error initializing sample data: {e}")

# Import routes after app creation
import routes  # noqa: F401
