requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyPDF2==3.0.1
python-docx==1.2.0
spacy==3.7.2
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
Werkzeug==3.0.1
Flask-SocketIO==5.3.6
eventlet==0.33.3
psycopg2-binary==2.9.7
email-validator==2.0.0

openai==1.14.3  # (or latest stable that works)
httpx==0.27.0   # Must match OpenAI SDK compatibility
faiss-cpu==1.7.4
Flask
python-dotenv

